requirements-minimal.txt
numpy>=1.16.0
PyYAML>=5.3.1
protobuf>=3.12.0

